#1443 split tx/dispatch.rs into cohesive submodules
Pure code-motion split of the 1474-LOC tx/dispatch.rs into the
module-with-children layout used by #1166 (TSO), #1354/#1586
(transmit), and #1591 (drain). Mirror module is now:

    userspace-dp/src/afxdp/tx/dispatch/
    ├── mod.rs              (994 LOC — orchestrator
    │                        enqueue_pending_forwards + Phase 8
    │                        body + recycle_ingress_frame +
    │                        resolve_pending_forward_target_binding
    │                        + predicates + re-export contract)
    ├── cos.rs              (105 LOC — Phase 1 + CoS fast-path
    │                        helpers)
    ├── shared_recycle.rs   (206 LOC — Phase 10 + cross-tick
    │                        recycle routing + resolve_tx_binding_
    │                        ifindex)
    ├── slow_path.rs        (292 LOC — handle_forward_build_failure
    │                        + maybe_reinject_slow_path* +
    │                        extract_l3_packet*, marked #[cold]
    │                        #[inline(never)])
    └── dispatch_tests.rs   (verbatim move; #[path] resolves
                             relative to mod.rs)

Re-export contract preserves every pub(in crate::afxdp) symbol
verbatim so the `use self::tx::dispatch::*;` glob at
afxdp/mod.rs:140 and the 16+ external resolve_tx_binding_ifindex
callers see no API surface change. extract_l3_packet and
extract_l3_packet_from_frame keep their pre-split sibling-tx/
visibility via `pub(in crate::afxdp::tx)`.

Hot/cold inlining strategy per plan v3:
- cos.rs helpers: #[inline] (per-request hot, single caller)
- slow_path.rs reinjection family: #[cold] + #[inline(never)]
  per AGY round-2 finding D — #[cold] alone does not stop LLVM
  from inlining a single-caller helper and bloating the hot
  i-cache footprint; #[inline(never)] guarantees the cold body
  stays out-of-line.
- shared_recycle.rs helpers: no markers (called per-iteration
  but body loops over post_recycles — let LLVM see the body).
- extract_l3_packet*: no markers (called from both hot and
  cold paths).

Out of scope deviation from plan v3: the 845-LOC body of
enqueue_pending_forwards (Phase 8 try_inplace_rewrite_or_build,
the Phase8Outcome enum, the four mid-body `continue` paths at
dispatch.rs:544/562/822/840 documented in plan v3) stays in
mod.rs. The Phase 8 body extraction is deferred to a follow-up
so reviewers can compare control flow against current master
without a body-shape diff. mod.rs at 994 LOC is below the
1000-LOC engineering-style cue, so the structural win is
sufficient for #1443; the Phase 8 body extraction trades risk
against modest additional cohesion gain and is better as its
own scoped PR.

Plan + adversarial review:
- Plan doc: docs/pr/1443-tx-dispatch-modularize/plan.md
- Codex R1: PLAN-NEEDS-MAJOR (task-mpn9fpo5-h6c0nu) — 5
  blockers, all addressed in plan v2.
- AGY R1: PLAN-NEEDS-MAJOR (adversarial-review-mpn90zjs-vntq8b)
  — 4 blockers, all addressed in plan v2.
- AGY R2: PLAN-NEEDS-MINOR (adversarial-review-mpna1lxa-o4ddp4)
  — 4 findings, all addressed in plan v3.
- AGY R3: PLAN-READY with quote-line verification
  (adversarial-review-mpnavi4h-1rqg68).
- Codex R2 + R3: infra-blocked across 4 attempts (codex-linux-
  sandbox ENOENT). Per [[feedback_codex_infra_must_retry]]
  retries dispatched; persistent infra outage. Codex hostile
  code review on the PR diff is the second-reviewer
  verification path.

Test plan:
- cargo build --release clean.
- cargo test --release: 1506 tests pass (1436 + 46 + 8 + 16).
  The pre-existing master-broken snat_contract_doc_guard test
  remains the sole failure (docs/userspace-dataplane-gaps.md
  predates the doc requirement on master).
- 5x flake check on
  enqueue_pending_forwards_mirrors_live_frame_and_records_counter:
  5/5 pass.
- Go test suite (`go test ./...`): all 30 packages green.

Closes #1443. Refs #1016 (#1016's outstanding "decouple
mutation from dispatch upstream" half is gated on #946 / #963
target architecture and explicitly out of scope).
